remove exchanges() API endpoint
This method does not currently accept pagination options, but it really
should.  The underlying rabbitmq management API has continuation support
landed, but not in a released version and not in a version of RbabitMQ
we are running.  Providing the endpoint without pagination would lead to
a breaking API change later.  Emulating pagination is possible, but we
currently do not have an anticipated use for this endpoint and it
returns a lot of useless information (test exchanges, etc.), so that
effort will not be worthwhile.  So, just delete it.